workaround around nosyntax problem on new buffer
using nosyntax produces a bug for me when having

call SyntaxRange#Include('```nosyntax'         ,'```'       ,'nosyntax'       ,'NonText')

in my .vimrc when openning subsequent buffer (not the first one, but the
second and so on, where :syntax is left at "off").

Adding this dummy, empty, syntax provides an alternative which works
out for me.